Finished functionality of updating draft general test questions Renamed db models for answers type specific info Minor bug fixes

+++ vokimi_api/Src/dtos/shared/general_test_creation/draft_general_test_answers/DraftGeneralTestImageOnlyAnswerFormData.cs
@@ -8,7 +8,7 @@ namespace vokimi_api.Src.dtos.shared.general_test_creation.draft_general_test_an
         public DraftGeneralTestImageOnlyAnswerFormData(
             string image,
             Dictionary<DraftGeneralTestResultId, string> relatedResultsIdName,
-            ushort orderInQuestion
+            int orderInQuestion
         ) {
             Image = image;
             RelatedResultsIdName = relatedResultsIdName;
